01-docker-terraform/requirements.txt
greenlet==3.1.1
numpy==2.2.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sqlalchemy==2.0.37
typing-extensions==4.12.2
tzdata==2024.2
